Home page and extra components
The home page contains navbar, searchboxes, sign up and log in forms, top suggestions. The validation of inputs in sign up form was added.
All the components are responsive.
The app was created using ReactJS, it can be run by changing path to frontend folder and executing command - npm start.